+++ flink-end-to-end-tests/flink-datastream-allround-test/src/main/resources/log4j.properties
@@ -16,8 +16,6 @@
 # limitations under the License.
 ################################################################################
 
-log4j.rootLogger=OFF, testlogger
-
 log4j.appender.testlogger=org.apache.log4j.ConsoleAppender
 log4j.appender.testlogger.target = System.err
 log4j.appender.testlogger.layout=org.apache.log4j.PatternLayout
