+++ flink-runtime/src/test/java/org/apache/flink/runtime/jobmanager/scheduler/SchedulerIsolatedTasksTest.java
@@ -26,7 +26,6 @@ import org.apache.flink.runtime.jobmaster.LogicalSlot;
 import org.apache.flink.runtime.taskmanager.TaskManagerLocation;
 import org.apache.flink.runtime.testingUtils.TestingUtils;
 
-import org.junit.Ignore;
 import org.junit.Test;
 
 import java.util.ArrayList;
@@ -113,15 +112,11 @@ public class SchedulerIsolatedTasksTest extends SchedulerTestBase {
 		assertEquals(5, testingSlotProvider.getNumberOfAvailableSlots());
 	}
 
-	@Ignore
 	@Test
 	public void testScheduleQueueing() throws Exception {
 		final int NUM_INSTANCES = 50;
 		final int NUM_SLOTS_PER_INSTANCE = 3;
 		final int NUM_TASKS_TO_SCHEDULE = 2000;
-
-		// note: since this test asynchronously releases slots, the executor needs release workers.
-		// doing the release call synchronous can lead to a deadlock
 
 		for (int i = 0; i < NUM_INSTANCES; i++) {
 			testingSlotProvider.addTaskManager((int) (Math.random() * NUM_SLOTS_PER_INSTANCE) + 1);
@@ -138,36 +133,6 @@ public class SchedulerIsolatedTasksTest extends SchedulerTestBase {
 		// flag to track errors in the concurrent thread
 		final AtomicBoolean errored = new AtomicBoolean(false);
 
-		// thread to asynchronously release slots
-		Runnable disposer = new Runnable() {
-
-			@Override
-			public void run() {
-				try {
-					int recycled = 0;
-					while (recycled < NUM_TASKS_TO_SCHEDULE) {
-						synchronized (toRelease) {
-							while (toRelease.isEmpty()) {
-								toRelease.wait();
-							}
-
-							Iterator<LogicalSlot> iter = toRelease.iterator();
-							LogicalSlot next = iter.next();
-							iter.remove();
-
-							next.releaseSlot();
-							recycled++;
-						}
-					}
-				} catch (Throwable t) {
-					errored.set(true);
-				}
-			}
-		};
-
-		Thread disposeThread = new Thread(disposer);
-		disposeThread.start();
-
 		for (int i = 0; i < NUM_TASKS_TO_SCHEDULE; i++) {
 			CompletableFuture<LogicalSlot> future = testingSlotProvider.allocateSlot(new ScheduledUnit(getDummyTask()), true, SlotProfile.noRequirements(), TestingUtils.infiniteTime());
 			future.thenAcceptAsync(
@@ -181,7 +146,25 @@ public class SchedulerIsolatedTasksTest extends SchedulerTestBase {
 			allAllocatedSlots.add(future);
 		}
 
-		disposeThread.join();
+		try {
+			int recycled = 0;
+			while (recycled < NUM_TASKS_TO_SCHEDULE) {
+				synchronized (toRelease) {
+					while (toRelease.isEmpty()) {
+						toRelease.wait();
+					}
+
+					Iterator<LogicalSlot> iter = toRelease.iterator();
+					LogicalSlot next = iter.next();
+					iter.remove();
+
+					next.releaseSlot();
+					recycled++;
+				}
+			}
+		} catch (Throwable t) {
+			errored.set(true);
+		}
 
 		assertFalse("The slot releasing thread caused an error.", errored.get());
