+++ flink-runtime/src/test/java/org/apache/flink/runtime/jobmanager/slots/DummySlotOwner.java
@@ -26,6 +26,6 @@ import org.apache.flink.runtime.jobmaster.SlotOwner;
  */
 public class DummySlotOwner implements SlotOwner {
 	@Override
-	public void returnAllocatedSlot(LogicalSlot logicalSlot) {
+	public void returnLogicalSlot(LogicalSlot logicalSlot) {
 	}
 }
