+++ flink-runtime/src/test/java/org/apache/flink/runtime/jobmaster/slotpool/AvailableSlotsTest.java
@@ -21,7 +21,6 @@ package org.apache.flink.runtime.jobmaster.slotpool;
 import org.apache.flink.runtime.clusterframework.types.AllocationID;
 import org.apache.flink.runtime.clusterframework.types.ResourceID;
 import org.apache.flink.runtime.clusterframework.types.ResourceProfile;
-import org.apache.flink.runtime.clusterframework.types.SlotProfile;
 import org.apache.flink.runtime.jobmanager.slots.TaskManagerGateway;
 import org.apache.flink.runtime.taskmanager.TaskManagerLocation;
 import org.apache.flink.util.TestLogger;
@@ -30,7 +29,6 @@ import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -80,28 +78,6 @@ public class AvailableSlotsTest extends TestLogger {
 		assertFalse(availableSlots.contains(slot3.getAllocationId()));
 		assertFalse(availableSlots.containsTaskManager(resource1));
 		assertFalse(availableSlots.containsTaskManager(resource2));
-	}
-
-	@Test
-	public void testPollFreeSlot() {
-		SlotPoolImpl.AvailableSlots availableSlots = new SlotPoolImpl.AvailableSlots();
-
-		final ResourceID resource1 = new ResourceID("resource1");
-		final AllocatedSlot slot1 = createAllocatedSlot(resource1);
-
-		availableSlots.add(slot1, 1L);
-
-		assertEquals(1, availableSlots.size());
-		assertTrue(availableSlots.contains(slot1.getAllocationId()));
-		assertTrue(availableSlots.containsTaskManager(resource1));
-
-		assertNull(availableSlots.poll(LocationPreferenceSchedulingStrategy.getInstance(), SlotProfile.noLocality(DEFAULT_TESTING_BIG_PROFILE)));
-
-		SlotAndLocality slotAndLocality = availableSlots.poll(LocationPreferenceSchedulingStrategy.getInstance(), SlotProfile.noLocality(DEFAULT_TESTING_PROFILE));
-		assertEquals(slot1, slotAndLocality.getSlot());
-		assertEquals(0, availableSlots.size());
-		assertFalse(availableSlots.contains(slot1.getAllocationId()));
-		assertFalse(availableSlots.containsTaskManager(resource1));
 	}
 
 	static AllocatedSlot createAllocatedSlot(final ResourceID resourceId) {
